Command reorg pt4
