+++ src/game/ChatCommands/TeleportationAndPositionCommands.cpp
@@ -75,6 +75,7 @@ static char const* const areatriggerKeys[] =
     NULL
 };
 
+
 bool ChatHandler::HandleGoHelper(Player* player, uint32 mapid, float x, float y, float const* zPtr, float const* ortPtr)
 {
     float z;
