+++ src/game/WorldHandlers/Chat.cpp
@@ -3841,6 +3841,199 @@ void ChatHandler::BuildChatPacket(WorldPacket& data, ChatMsg msgtype, char const
     }
 }
 
+// Shared Helpers between command implementation files
+
+void ChatHandler::ShowFactionListHelper(FactionEntry const* factionEntry, LocaleConstant loc, FactionState const* repState /*= NULL*/, Player* target /*= NULL */)
+{
+    std::string name = factionEntry->name[loc];
+    // send faction in "id - [faction] rank reputation [visible] [at war] [own team] [unknown] [invisible] [inactive]" format
+    // or              "id - [faction] [no reputation]" format
+    std::ostringstream ss;
+    if (m_session)
+    {
+        ss << factionEntry->ID << " - |cffffffff|Hfaction:" << factionEntry->ID << "|h[" << name << " " << localeNames[loc] << "]|h|r";
+    }
+    else
+    {
+        ss << factionEntry->ID << " - " << name << " " << localeNames[loc];
+    }
+
+    if (repState)                               // and then target!=NULL also
+    {
+        ReputationRank rank = target->GetReputationMgr().GetRank(factionEntry);
+        std::string rankName = GetMangosString(ReputationRankStrIndex[rank]);
+
+        ss << " " << rankName << "|h|r (" << target->GetReputationMgr().GetReputation(factionEntry) << ")";
+
+        if (repState->Flags & FACTION_FLAG_VISIBLE)
+        {
+            ss << GetMangosString(LANG_FACTION_VISIBLE);
+        }
+        if (repState->Flags & FACTION_FLAG_AT_WAR)
+        {
+            ss << GetMangosString(LANG_FACTION_ATWAR);
+        }
+        if (repState->Flags & FACTION_FLAG_PEACE_FORCED)
+        {
+            ss << GetMangosString(LANG_FACTION_PEACE_FORCED);
+        }
+        if (repState->Flags & FACTION_FLAG_HIDDEN)
+        {
+            ss << GetMangosString(LANG_FACTION_HIDDEN);
+        }
+        if (repState->Flags & FACTION_FLAG_INVISIBLE_FORCED)
+        {
+            ss << GetMangosString(LANG_FACTION_INVISIBLE_FORCED);
+        }
+        if (repState->Flags & FACTION_FLAG_INACTIVE)
+        {
+            ss << GetMangosString(LANG_FACTION_INACTIVE);
+        }
+    }
+    else if (target)
+    {
+        ss << GetMangosString(LANG_FACTION_NOREPUTATION);
+    }
+
+    SendSysMessage(ss.str().c_str());
+}
+void ChatHandler::ShowSpellListHelper(Player* target, SpellEntry const* spellInfo, LocaleConstant loc)
+{
+    uint32 id = spellInfo->Id;
+
+    bool known = target && target->HasSpell(id);
+    bool learn = (spellInfo->Effect[EFFECT_INDEX_0] == SPELL_EFFECT_LEARN_SPELL);
+
+    uint32 talentCost = GetTalentSpellCost(id);
+
+    bool talent = (talentCost > 0);
+    bool passive = IsPassiveSpell(spellInfo);
+    bool active = target && target->HasAura(id);
+
+    // unit32 used to prevent interpreting uint8 as char at output
+    // find rank of learned spell for learning spell, or talent rank
+    uint32 rank = talentCost ? talentCost : sSpellMgr.GetSpellRank(learn ? spellInfo->EffectTriggerSpell[EFFECT_INDEX_0] : id);
+
+    // send spell in "id - [name, rank N] [talent] [passive] [learn] [known]" format
+    std::ostringstream ss;
+    if (m_session)
+    {
+        ss << id << " - |cffffffff|Hspell:" << id << "|h[" << spellInfo->SpellName[loc];
+    }
+    else
+    {
+        ss << id << " - " << spellInfo->SpellName[loc];
+    }
+
+    // include rank in link name
+    if (rank)
+    {
+        ss << GetMangosString(LANG_SPELL_RANK) << rank;
+    }
+
+    if (m_session)
+    {
+        ss << " " << localeNames[loc] << "]|h|r";
+    }
+    else
+    {
+        ss << " " << localeNames[loc];
+    }
+
+    if (talent)
+    {
+        ss << GetMangosString(LANG_TALENT);
+    }
+    if (passive)
+    {
+        ss << GetMangosString(LANG_PASSIVE);
+    }
+    if (learn)
+    {
+        ss << GetMangosString(LANG_LEARN);
+    }
+    if (known)
+    {
+        ss << GetMangosString(LANG_KNOWN);
+    }
+    if (active)
+    {
+        ss << GetMangosString(LANG_ACTIVE);
+    }
+
+    SendSysMessage(ss.str().c_str());
+}
+
+bool ChatHandler::ShowPlayerListHelper(QueryResult* result, uint32* limit, bool title, bool error)
+{
+    if (!result)
+    {
+        if (error)
+        {
+            PSendSysMessage(LANG_NO_PLAYERS_FOUND);
+            SetSentErrorMessage(true);
+        }
+        return false;
+    }
+
+    if (!m_session && title)
+    {
+        SendSysMessage(LANG_CHARACTERS_LIST_BAR);
+        SendSysMessage(LANG_CHARACTERS_LIST_HEADER);
+        SendSysMessage(LANG_CHARACTERS_LIST_BAR);
+    }
+
+    if (result)
+    {
+        ///- Circle through them. Display username and GM level
+        do
+        {
+            // check limit
+            if (limit)
+            {
+                if (*limit == 0)
+                {
+                    break;
+                }
+                --* limit;
+            }
+
+            Field* fields = result->Fetch();
+            uint32 guid = fields[0].GetUInt32();
+            std::string name = fields[1].GetCppString();
+            uint8 race = fields[2].GetUInt8();
+            uint8 class_ = fields[3].GetUInt8();
+            uint32 level = fields[4].GetUInt32();
+
+            ChrRacesEntry const* raceEntry = sChrRacesStore.LookupEntry(race);
+            ChrClassesEntry const* classEntry = sChrClassesStore.LookupEntry(class_);
+
+            char const* race_name = raceEntry ? raceEntry->name[GetSessionDbcLocale()] : "<?>";
+            char const* class_name = classEntry ? classEntry->name[GetSessionDbcLocale()] : "<?>";
+
+            if (!m_session)
+            {
+                PSendSysMessage(LANG_CHARACTERS_LIST_LINE_CONSOLE, guid, name.c_str(), race_name, class_name, level);
+            }
+            else
+            {
+                PSendSysMessage(LANG_CHARACTERS_LIST_LINE_CHAT, guid, name.c_str(), name.c_str(), race_name, class_name, level);
+            }
+        } while (result->NextRow());
+
+        delete result;
+    }
+
+    if (!m_session)
+    {
+        SendSysMessage(LANG_CHARACTERS_LIST_BAR);
+    }
+
+    return true;
+}
+
+
+
 // Instantiate template for helper function
 template void ChatHandler::ShowNpcOrGoSpawnInformation<Creature>(uint32 guid);
 template void ChatHandler::ShowNpcOrGoSpawnInformation<GameObject>(uint32 guid);
